Smarter caching in MythPainter.
Previously we attempted to keep a check on memory consumption by
tracking the total number of hardware buffers created for OpenGL,
Direct3D9 and VDPAU painting and releasing resources when the count
exceeded 256. A little debugging showed that it was fairly easy to reach
over 300Mb of allocate memory, which was confirmed via OpenGL profiling.
Given that OpenGL will mirror hardware buffers in main memory, that
amounts to a sizeable chunk of memory consumption even by modern
standards.

This is now replaced with an estimate of the memory consumed by the
bitmaps and begins to release resources above a consistent limit. I've
set this at 64Mb, which appears to work without issue on even the most
resource hungry themes and should represent around 8 full screens of
bitmaps when displayed at 1920x1080. It can of course be tweaked if it
proves to be inadequate.

There is a also a knock on benefit for VDPAU playback on graphics cards
with 256Mb of memory. As OpenGL will not immediately release memory to
other 'services' even when released by the application, this approach to
caching should limit the land grab made by the OpenGL element of the
driver.

I've not touched the Qt painter as it doesn't seem to have any cache
management of its own. Which is odd.


git-svn-id: http://svn.mythtv.org/svn/trunk@26085 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythui/mythpainter_ogl.h
@@ -60,12 +60,12 @@ class MPUBLIC MythOpenGLPainter : public MythPainter
     int               target;
     bool              swapControl;
 
-    QMap<MythImage *, unsigned int> m_ImageIntMap;
-    std::list<MythImage *>          m_ImageExpireList;
-    QMap<QString, MythImage *>      m_StringToImageMap;
-    std::list<QString>              m_StringExpireList;
-    std::list<long long>            m_textureDeleteList;
-    QMutex                          m_textureDeleteLock;
+    QMap<MythImage *, uint>    m_ImageIntMap;
+    std::list<MythImage *>     m_ImageExpireList;
+    QMap<QString, MythImage *> m_StringToImageMap;
+    std::list<QString>         m_StringExpireList;
+    std::list<uint>            m_textureDeleteList;
+    QMutex                     m_textureDeleteLock;
 };
 
 #endif
